cube-ctl: Fix state problem where container cannot be restarted
Sometimes when the container is stopped the delete operation cannot be
processed immediately because the container state updates are not
fully synced yet.  Example:

    root@cube-dom0:/usr/sbin# c3 stop cube-desktop
    [INFO] stopping container cube-desktop
    cannot delete container cube-desktop that is not stopped: running

The latest oci runtime now allows the delete operation to forcefully
kill the container so that it can be terminated in one step reliably,
so we will switch to using it instead of implementing a check for the
state to change prior to initiating the delete.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>