cube-ctl: Fix state problem where container cannot be restarted
Sometimes when the container is stopped the delete operation cannot be
processed immediately because the container state updates are not
fully synced yet.  Example:

    root@cube-dom0:/usr/sbin# c3 stop cube-desktop
    [INFO] stopping container cube-desktop
    cannot delete container cube-desktop that is not stopped: running

The latest oci runtime has added a -f argument to the delete command
which will send a signal and wait.  In the case that signal was
already received it still waits until the conatiner is actually
stopped and then purges the configuration.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>