cube-cfg: variable specific set hooks
Allow a defined variable to have a hook that is called when the value
is changed.

This allows us to trigger more complex operations when a value is
changed.

The first user of this is the app variable which adds/removes
systemd-hooks if the app entry point is /sbin/init.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>