cube-cfg: change 'name' to 'port'
Unfortunately the original use of 'name' was done as part of a larger
piece of work so there is no record if this was intentional or not. In
the script this is used as a 'port' and I can't determine any
intention for this to be something other than a port, so changing
'name' to 'port' in the help text.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>