pflask: switch to internal hooks
pflask has no native hook support, so it was approximated by calling
a series of scripts after pflask has started a container.

This unfortunately leads to race conditions as the entry point to
the container has started running, while setup and other hook activities
are also executing.

Things like network interfaces, etc, may or may not be setup in time
for the hooks .. and hence inconsistent results.

With this change, we add a --hook parameters to pflask, which will
executed any number of hooks before starting the main container
entry point.

Each hooks is called with:

  <hook> <container directory> <container entry point pid>

If a hook is specified as:

 <hook> <hook arguments> %s

The system arguments of <container directory> <container entry point pid>
will replace %s before calling the hook.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>