overc-utils: add bash completions file
Using the help text at the top of cube-ctl and cube-cfg as a
'template' this adds intelligent bash completion when using the 'c3',
'c3-cfg', 'c3-ctl', 'cube-cfg' or 'cube-ctl' commands. As with any
bash completions these are only activated with mutiple presses of the
tab key, so if you like to type c3 commands in full you will still be
able to. For people using the completions they will be presented with
the available options when considerring what is already present on the
cmdline. Changes to the assiciated scripts should be accompanied by
changes to the completions, as needed.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>