cube-cfg: Fix default startup of cgroup isolation
When the OCI containers start from dom0 via the systemd rule, the
systemd from dom0 will nest the cgroup of the OCI container.  The
behavior changed recently with systemd when it started to make use of
the cgroup2 functionality.

There are three parts to the fix:

1) Each new OCI container should create a cgroup2 controller and the
   startup script should reparent itself into the script.

2) The config.json generation should specify the cgroup location else
   it will inherit what ever systemd assigns.

3) The stdin/stdout/stderr need to be closed for any forked process in
   the container startup script else devices from dom0 (which performs
   the startup from systemd) will not be released until that process
   is killed.

The internal test of the plumbing to ensure dom0 is properly separated
from the OCI containers is to boot the system and from cube-essential
run the following:

    # Stop dom0 forcefully
    systemctl stop essential-autostart.service
    # Check dom0 is gone but the other container are still around
    machinectl
    /var/lib/cube/essential/sbin/runc list
    # Restart dom0
    systemctl start essential-autostart.service

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>