pflask: Fix occasional infinite loop on console exit
Depending on how a pflask exits and if it had an attached console
client, the client occasionally gets into a state where it will loop
trying to execute a read that will never succeed because the server
side is already gone.

The "kill 0" system call can be used to check if the service side
process is still alive without actually sending a signal and allow the
console code to exit gracefully.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>