cube-cfg: add container environment variable support and oci.cmd cleanup
Add the ability to store container environment variables in cube
configurations.

To implement this, we add a new flag to the get/set variables of
'type'. The type dictates how we'll set/get values from the variable.
For environment variables, we create a 'multi' type, which stores
multiple key/pair values.

Example environment value setting:

   cube-cfg -n mycontainer set cube.env:PATH=/bin/

We also change the generated oci.cmd from a set of echo statements
to a HEREDOC. This makes the script easier to maintain in cube-cfg.

As part of this change, we make more elements of the oci.cmd
check values at runtime (rather than being mainly statically
generated), so we can switch elements like console/pty allocation
at runtime.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>