cube-ctl: foreground containers and environment variables
To enhance support for oci/docker containers, we add two features
in this commit: foreground container support and env vars.

foreground containers are triggered with the --foreground or -f
flag to c3. This in turn changes the pty allocation and config.json
to output directly to the running console. This allows the debug
of container startup issues.

  % c3 start -f mycontainer

Note: if your container daemonizes or runs in the background, only
      use this for debugging.

The second change in this commit is the import of environment
variables from oci/docker config.json files. Many docker containers
not only have an entry point, but specify a set of necessary
environment variables for that application.

With this change, our imported containers will pick up these
values and make them persistent if the container is then stored
to a docker registry.

 % c3 add docker://httpd
 % c3 start httpd
 % c3-cfg auth docker:<username>:<password>
 % c3 store -n httpd docker://<username>/httpd

We also have a small fix to distinguish between containers
of similar names (i.e. httpd and httpd-new are now not confused
due to bad regex matching).

Finally we do a force stop on containers that are being
stopped to ensure that resources are always freed.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>