cube-ctl: add initial support for tags
Previously all operations on oci or docker containers enforced the 'latest'
tag.

With this commit, we can now vary the tag on both the source and destination
using the OCI syntax of :<tag> following the source or dest.

There are still restrictions, like the source and dest must be the same
tag (and that may stay), but it is otherwise functional.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>