cube-ctl: make container name contruction smarter
If someone forgets to add the -n <name> to an add command, the
name is generated. But with docker containers, they can get invalid
pretty fast.

So we make the naming a bit smarter and strip tags, ports, and
convert paths appropriately.

We also add a --track option (currently hidden), since the container
tracking is throwing and error that we don't want to see.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>