cube-*: add 'runonce' capability to settings
Add the ability to set a .runonce value for the various settings.

These settings will be prioritized over the existing settings and
will be cleared after a single use.

If runonce values are found, the .json and other settings are
modified for the run, and then restored to their defaults after
(by removing the .runonce settings and regenerating).

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>