overc-system-agent: overc: support both live output and normal output
In most cases we need live output from the template scripts, but it's
not always in that situation, for instance, when we list containers, we
do not need that, it breaks the output and receives incomplete lines
for some unknown reason. So we introduce a option 'liveoutput' to
control if we want live output or not in a command run.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>