qemu: setup proper PACKAGECONFIG to make things useful
There is a number of different ways to set these PACKAGECONFIGS and in
a number of locations. We can do this in meta-virt or meta-overc for
example. We can overwrite the PACKAGECONFIG from qemu.inc using '?='
which allows others to overwrite things in their own layers. Or we can
use this method to extend what is set in qemu.inc. Choices and more
choices. I think we should have something in meta-virt which is
overwriteable. For now I would recommend this be added to meta-overc
as we can have the last word there, without the requirement (or
desire) to let others overwrite our choices.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>