networking: initiate .cube.lan
Since not all containers are avahi enabled, it means that ips must
be used when navigating within the platform.

dnsmasq can help us to avoid this, but creating a local .cube.lan
domain, and expanding dns queries that match.

The vrf sets this up in dnsmasq and our networking hooks put the
name of the container and their IPs into the /etc/hosts of the
VRF.

Note: the entry of the container names into the /etc/hosts could be
arbirated, but so far, there are no issues with normal file access
(since order doesn't matter).

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>