cube-cmd/cube-ctl: Allow start -f to be interactive or non-interactive
The -f option is ONLY for debugging the startup of a container.

It has two ways to use it.  For a daemon container which will never
use stdin you would use:

c3 start -f cube-X

For an interactive container which has a shell or login prompt you
would run:

c3 start -i -f cube-X

In both cases all of the console integration is disabled and the
config.json terminal attribute is the only difference.  The
stdin/stdout/stderr as well as the terminal controls for signaling
like "control-c" will be pass directly to the container from the shell
you run "c3 start" in.  In order to obtain the controlling pty of the
shell that invoked cube-ctl, the subshell parentheses were removed.

This now means you can actually execute a control-c in a
non-interactive container like the httpd.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>