cube-ctl/cube-cfg: Add c3 start -a cube-X
For OCI containers add an option to start to immediately attach to the
console.  You will be able to detach per the normal dtach detach
sequence while not affecting any other part of the container start.

This option is useful for debugging the container start when using
either the tty or screen-tty console manager.

When using c3 start -a, oci.cmd is actually run twice, with a command
in between:
   1) oci.cmd consoleonly
      - This will startup the console hanlding code for tty or screen-tty mode
      - Then it exits before calling the container startup

   2) the cube-ctl runs dtach -a /var/lib/cube/${container}/dtach-sock

   3) oci.cmd attach
      - This time the console handling startup is skipped
      - The container startup code is executed

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>