essential-init: add reload-dom0-snapshot.service
This service will handle reloading dom0 snapshot if it's detected that
dom0 had been rolled back at the last running time, this service is
needed because dom0 can not reload itself.

When this service starts, it will check a file existence:
/opt/container/.snapshots/dom0/.need_reload

and read the snapshot ID from the file if it exists, or else it will do
nothing.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>