cube-cfg/ctl: fix linkage and container manager
The linkage of a container a -> b means that rules will be executed
in container a .. those rules need to know the IP address of b to
work properly. As such, if you link two containers, we throw the
autostart flag to make sure it runs on the next boot.

We also add an extra check to ensure that empty IP addresses do not
trigger bad rules.

And finally when switching to runv, we need to use the container
name or the configuration value won't stick.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>