overc-utils: remove overc-installer git repo from SRC_URI
The cubename is the only script that is being required by overc-utils
recipe(others are all local scripts), we can move it to overc-installer
recipe since it's the correct place to install cubename, so we do not
have to maintain a same repo in multiple recipes, which is not
necessary and not reasonable as well from my point of view.

After the change, overc-utils needs set RDEPENDS to overc-installer,
and change the PV to 0.1 since it's not a git recipe any more.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>