cube-ctl: normalize container status/lists/processing
The listing, manipulation and checking of container states grew
organically over time. In order to support new container managers
(i.e. runv, lxc) the routines need to be more general and use
utility functions (versus going directly at the commands).

This commit enforces the separation of manager -> commands, and
also re-works the container lists to be consistent.

We also introduce the ability to sort by type, name or state.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>