pflask: Use RegisterMachine instead of CreateMachine
The systemd machined was intended for use with systemd-nspawn when
using the RegisterMachine interface.  The problem is that this causes
the dom0 pflask to get its cgroups managed by systemd and the signals
that are passed to the systemd running in dom0.  The dom0 is intended
to manage the system and as a result it causes problems with circular
signaling that causes hangs on reboot and halt operations.

Systemd also provides a RegisterMachine interface which is more well
suited for the problem.  It will use only the PID management features
instead of trying to control the cgroups and issue signals for the
various shutdown use cases.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>