container-shutdown-notifier: Fix systemd dependency to root ns
There are two problems to fix here to allow reboot/shutdown to not
hang when launched from a container.

1) The reboot command (which comes from systemctl) in the parent name
space is partially synchronous, and will block when launched via
cube-cmd, specifically in the dom0 case.  Instead we can work around
this by directly sending a signal to the init process and exit
immediately.

2) If a reboot is already in progress in the parent name space another
reboot additional reboot signaling can cause systemd to block for 1
min 30seconds before executing a series of kill operations, so we check
for this corner case.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>