essential-autostart: Issue graceful stop to dom0 by default
We want the following to be true when stopping dom0:

1) manually run: systemctl stop essential-autostart

   * This is debugging only it is a kill of dom0 so you can restart it
     without harming anything else that is running.

2) reboot/shutdown/halt from essential

   * Check if dom0 is still running and if not do nothing

   * The stop sequence should request dom0 to run its shutdown
     sequence and wait for it to complete (it gets 1min and 30 seconds
     as specified by systemd).

   * In a prior patch, dom0 breaks the circular signaling by checking
     if essential is executing the reboot.

The stop logic was moved into the same script that starts dom0.

The essential-autostart.service now depends on machines.target and
getty.target.  These targets were already implicitly started at
service start time, but we need the dependency for shutdown.  We need
to ensure the essential-autostart.service stops completely before the
machines.target else we cannot talk to machined.  We need the
getty.target to be run after the stop else we cannot debug any
problems because it will terminate the console I/O in parallel.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>