gtk-xfce-engines: fix utf compilation issues with glib 2.53.4
Reference: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=870738I

With the glib update in oe, the following error may be seen:

  UnicodeDecodeError: 'utf-8' codec can't decode byte 0xd6 in position 980: invalid continuation byte

The debian patch removes the offending character and we can build again.
This is done in a bbappend, since the error does not seem to be universally seen.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>