cube-cmd/essential-init: Add ability to daemonize runc
If runc has been modified to daemonize itself and properly run its
stop hooks upon termination of the runc process, it requires a helper
process which is nothing more than a fork and exec wrapper.

The wrapper function performs 3 simple tasks:

1) Set env variable SIGUSR1_PARENT_PID which will cause runc to signal
   back to the parent that it wants to daemonize.

2) Fork and exec a "runc start ..." command

3) Wait for the child process to send a SIGUSR1 or exit and then
   exit leaving the child process behind if it is a SIGUSR1

In the case that runc does not contain any modifications to deal with
the SIGUSR1, the wrapper automatically exits because the typical runc
start operation is designed to exit.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>