Revert "dom0: esssential: use shared linux firmware"
This reverts commit f8091aecadbf32b33cf5a2facf5902d5eacf0d54.

Even though the systemd-modules-load.service had been disabled
in essential, but the udev would trigge to load kernel modules, too.
Thus the firmwares should be kept in essential instead of in dom0,
otherwise, some kernel moduels will be loaded in essential but failed
to load the firmwares, such as iwlwifi module failed to load
the firmware iwlwifi-7260-17.ucode.