nctl: make nanomsg read/response predicable and fix ACLs
When nanomsg reads from ipc:// the message will continue to be
read until a response is written.

At times, we were going back to the read loop before the response
was sent, as such, repeated commands were picked up.

We latch our last read command, and don't read any new ones until
a reply has been sent.

If we do try and read before the reply is sent, we output a warning.
The warning will allow future debug if races to re-appear.

We also cleanup/fix the ACL processing to understand compound
commands and not allow circumvention of the checking. This is done
by three checks:

  - do not allow background comamnds "&"
  - split on ; && || and do ACL checks on each part
  - construct a restricted shell enviroment to prep for unpriv commands

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>