cube-ctl: add remote management controls
The ability to add capabilities (including rootfs transfer) were broken
with some changes to the nctl client->server model.

To assist in accessing the capabilities interface, the c3-cap script is
provided, and could be used as follows:

  c3-cap --cap add alpine.tar.gz

And the alpine.tar.gz would be streamed to the monitoring container. You
could then install the container via:

  c3 add -t oci -n alpine /var/lib/cube-cmd-server/functions.d/alpine.tar.gz

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>