c3-ctl: source /etc/c3-ctl for defaults
If a layer wants to have a different set of defaults for c3 options
(i.e. container tracking), then they can create a file /etc/c3-ctl.rc
and supply bash variable defaults to c3.

Note: this is not expected to be a end user visible file, but is
provided to allow images to be tailored with different defaults and
to avoid requiring users to pass options to each call.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>