cube-ctl: fix the issue of tracking the docker container twice
In cube-ctl, there were two places doing the track action, one
was in the common place, and another one was properly for docker
container. Since the action of adding a docker container will
recursively call cube-ctl, thus it will track it twice. To fix
this issue, remove the track action specific to docker and let
it used the common place code.

Signed-off-by: Fupan Li <fupan.li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>