cube-install: initial version
Up to now we have used cube-essential as the rootfs in an OverC
installer image (live+payload). This worked well since cube-essential
was designed to run as a system rootfs and it avoided us having to
maintain another image type. Since we have moved to slim
cube-essential, however, it no longer contains all the applications
required to run the installer. We therefor add a new image type called
'cube-install'.

The intention of cube-install is initially to have a rootfs which can
be used in 'installer' images (live+payload). It contains everything
needed to boot plus the applications used by the installer
scripts. This is what is available in this commit.

In the future we should be able to add the ability to run cube-install
as a container. This will allow us for example to have OverC run from
a USB stick, created using the 'image' install, and then perform an
install to the system HDD by doing a 'c3 add cube-install', and
running the installer from cube-install. By removing the installation
binding from cube-essential and instead having a dedicated
cube-install we open the door to these possibilities, without
compromising cube-essential.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>