cube-console: error if container name is not recogized
When running cube-console on dom0, an unknown container was dropping
to cube-essential's namespace shell. This isn't expected, so instead
if we don't find a container manager, we error and indicate that the
container name is not known.

Also, we make the regex matching for dom0 and essential to be strict
to prevent typo's from attempting a shell to a container.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>