cube-desktop: don't include packagegroup-dom0
The packagegroup-dom0 is intended to be used only when creating the
cube-dom0. It includes things like linux-firmware which should not be
included in other images. If an image needs elements of
packagegroup-dom0 it is simple enough to add them as these components
are available via OVERC_COMMON_* and friends.

I don't believe there was every an active decision to include
packagegroup-dom0 (at least there is no specific evidence which is
easily found). It appears to come from commit 6693A483723D [default
packages: Add consul and etcd to the base overc image]. Without any
hints or pointers to specific usecases that this change might effect I
was only able to run our basic tests and ensure nothing broke.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>