c3-cfg: track whether or not a container needs a regen
For operations that modify the settings of a container, it is useful
to indicate in the container listing if the container should be
regenerated.

We can do this via a "dirty" field in the attributes.

Operations that modify a config, set a flag which triggers the attribute
and a regen of a container's configuration will clear the flag.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>