cube-ctl: fixup netprime detection when it is dom0
The current implementation of container_by_attribute() fails when
there are multiple attributes, such is the case for dom0 when it is
the 'vrf' and the 'netprime'. This can be seen with dom0 is both vrf
and netprime and 'c3 ctl netprime' is run, it will return an empty
value.

In container_by_attribute() the call to:
cube-cfg -n ${c} attribute

returns "vrf netprime" which will then fail the string comparison.

We could fix this in a few ways, the way I have it here or adding a
way to fetch individual attributes to cube-cfg. (ie. do a get when
neither '-' or '+' is passed). I opted for the fix included here since
all other code treats these attributes as one.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>