cube-ctl: use container_by_attribute() where possible
There are places where we had code doing the same work as
container_by_attribute(). This is prone to bugs and makes any future
fixes have to be applied in several places. So use
container_by_attribute() wherever possible.

NOTE: the code to move the netprime will only work on /opt/container
even though some of the code gives the impression it respects
out_dir. ie. some of the code respects out_dir, but not all.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>