c3 list: make container listing 'smarter'
It was possible to issue an invalid set command, and create a container
manager setting in the current working directory. This incorrectly placed
setting, could then be picked up by subsequent 'get' commands, which
allowed invalid containers to end up in the container listing.

To fix this, we tighten up the out_dir detection code to ensure that
/opt/container/<foo> actually exists, and also, to not care if config.json
exists.

Additionally, we also add a new container state of "invalid" if an
incorrect container manager is configured.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>