runv: update package lists to split between dom0 and essential
While we still need the support components of runv on essential
(qemu + bios + hyperv), we don't need the actual runv executable
on essential. In order to have a working runv container manager,
we need to add runv to dom0 when the "vm-sep" distro feature is
enabled.

We also teach cube-ctl to allow commands to be executed against
runv, by adding it to the known container manager case statments.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>