c3-app-container: Initial version of a basic app container
This adds a simple app container image which allows for the creation
of 1 app container image using only a local.conf variable, without the
need to touch any recipe files. As described in the included README
the use of an .inc file will allow users to create additional app
container images with only a few lines in a new image recipe file.

Eventually I would like to have a 'universal' app image which can
allow the building of any application (for which there is a recipe)
into its own container image, without the need to create image
recipes. This is the first (needed) step toward getting to this goal.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>