c3-systemd-container: initial version of a basic systemd system image
This adds a systemd system container image which allows for the
creation of 1 systemd system container image using local.conf
variable(s), without the need to touch any recipe files. As described
in the included README the use of an .inc file will allow users to
create additional systemd system container images with only a few
lines in a new image recipe file.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>