cube-cfg: need to strip "-" or "+" before determine_outdir()
Attempting to do something like
c3 cfg link -dom0:12345 desktop:12345

results in an error:
  grep: invalid argument 'eb' for '--directories'
  Valid arguments are:
  - 'read'
  - 'recurse'
  - 'skip'
  Usage: grep [OPTION]... PATTERN [FILE]...

This was a result of the '-' being passed and eventually used by grep
as if it marked an option. Since we don't use 'outdir' immediately
here we move the assignment below the stripping of the 'op' and add a
comment to hopefully prevent future code changes falling into the same
trap.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>