cube-ctl: catch all non-dashed, not just known or double-dashed
Attempting to do something like:
c3 add -d blah docker://debian

Will result in a fairly random error:
  basename: invalid option -- 'd'
  Try 'basename --help' for more information.

This is the result of '-d' being treated as a non-dashed option when
it is clearly dashed. We have code to catch unknown double-dash
options so extend this to also cover unknown single-dash options. This
will now catch the '-d' as an unknown option, as it should.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>