cube-cfg: mirror add when doing a remove
Everytime we currently do a remove of an attribute we are leaving an
extra space. This will be visible in the 'cube.attributes' file. For
example moving the netprime from and to dom0 several time will result
in the 'cube.attributes' file having increasingly more spaces between
'vrf' and 'netprime', each time:
  vrf netprime
  ...
  vrf  netprime
  ...
  vrf   netprime
  ...

The add always adds a new space so we now make the remove mirror this
and remove the extra space.

NOTE: Each space is interpretted as a ',' in 'c3 list' so without this
fix you end up seeing:

dom0 pflask running vrf,,,netprime 192.168.42.1,10.0.2.15

(with the # of ',' reflecting each move away and to dom0 of the
netprime)

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>