cube-ctl: add pflask specific remove for netprime poststart hook
The format of the .cube.hooks.poststart is different when comparing a
pflask managed container and an oci one. You can see this just a few
lines above this change when the netprime poststart hook is added.

Without this change if you were to move the netprime from dom0 to
another container the posstart hook would remain.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>