cube-ctl: don't match on substrings
Attempting to delete a container named 'deb' while a container named
'debian' is running will result in

  [INFO] container deb is running, it must be stopped before deleting

When comparing the container name to the list of running containers we
are matching when the container name is a substring of another
container in the list. We could make use of bash string matching to
avoid this but so far we have avoid using bash only constructs. We
could also grep on "($| )${container_name}( |$)" but the method used
here I believe reads the best.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>