c3-cfg: enable SYS_PTRACE and control ambient options by default
For the most part, we want to enable ptrace and control both
ambient and inheritable capabilities.

This may not be the case for some app container, but for now, we
control them via a flag ... but enable them all the time. We will
tweak things in the future when the use cases arrive.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>