ebtables: temporary fix for linux-libc-headers 4.15
There was a known bug included in linux-libc-headers 4.15 that was
reported and fixed in mainline. See
https://github.com/torvalds/linux/commit/da360299b673

The original breakage is part of the linux-libc-headers 4.15 which
causes ebtables to fail to build, with multiple redefinitions as
described in the fix commit.

The fix has been recommended for promotion to 4.15.y stable so we
should get a proper fix on the next linux-libc-headers uprev, but for
the time being we include a copy of the header with the offending
'include' removed to allow ebtables to build successfully.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>