c3-construct: tee output to screen by default and make full path available
When waiting for a long c3-construct, the user may wonder if things are
really working. So in addition to logging errors/output, we also tee them
both to the screen.

Also, we make a new variable PRIV_PATH available to scripts, so that they
can chose to pass the full path to some other commands (no path is the
default).

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>