c3-construct: fix stderr/stdout redirection, improve -v and add -w
The c3 construct script was not properly segmenting stderr and stdout,
so we fix the redirects.

It also could be very silent, so we automatically tee the output to the
screen, while also doing the redirects for forensics.

We also improve -v to do a 'set -x' so each command can be seen as it
is invoked.

And finally, we add -w so that a failure from a phase of the script
will not abort all processing.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>