cube-cfg: add the ability to link directories between containers
cube-cfg already had the ability to link the TCP/UDP ports of any
two containers. But is also a common need to share data stores between
containers.

This commit introduces a basic plumbing command to implement this
type of sharing. When a non-numeric link source/dest is detected, a
data link is triggered.

A data link can be a file, or a directory.

The steps to making the link work are:

  - The source container's directory structure is checked, and the
    source directory validated

  - If valid, a new export store is created in /opt/container/data/<container a>

  - The source directory is symlinked into that new data store

  - The destination container has a bind mount configured, and its configuration
    regenerated.

This is a plumbing command, and does very little error checking

Example:

  root@cube-dom0:~# c3-cfg -v link cadvisor:/etc debian:/tmp/blah
  [INFO]: linking cadvisor:/etc to debian:/tmp/blah

  root@cube-dom0:~# c3 start debian
  root@cube-dom0:~# c3 -i debian:/bin/bash
  root@debian:/# cat /tmp/blah/alpine-release
  3.4.6

  root@cube-dom0:~# c3 stop debian

  root@cube-dom0:~# c3-cfg -v link -cadvisor:/etc debian:/tmp/blah
  [INFO]: unlinking cadvisor:/etc from debian:/tmp/blah
  [INFO]: rm -rf /opt/container/data/cadvisor///etc

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>