cube-cfg: specify outdir when linking containers
When linking the ports of containers, we should explicitly pass the
output directory to the cube-cfg calls. We do this, since if running
in a chroot'd or mount'd scenario determine_outdir won't be able to
locate and automatically set the proper directory.

We also improve determine_outdir to always set a full path (versus
allowing ./), since we can use that full path to determine where
we are, and how to manipulate it for sub calls.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>